requirements.txt
Django==4.2.20
djangorestframework==3.14.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
drf-yasg==1.21.7
Faker==22.6.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.1 